+++ install_macos.sh
@@ -26,6 +26,7 @@ declare -a MACOS_MODULE_ORDER=(
     "github-cli"
     "gitlab-cli"
     "nvm"
+    "defguard"
 )
 
 declare -a MACOS_SKIP_MODULES=(